refactor: PostAdd/Edit/DetailDialog, UserDialog 분리
