+++ src/features/comment/ui/index.ts
@@ -1,4 +1,4 @@
-export * from "./AddCommentDialog";
-export * from "./EditCommentDialog";
-export * from "./CommentList";
-export * from "./CommentRow";
+export { default as CommentAddDialog } from "./CommentAddDialog";
+export { default as CommentEditDialog } from "./CommentEditDialog";
+export { default as CommentList } from "./CommentList";
+export { default as CommentRow } from "./CommentRow";
