+++ src/features/post/ui/index.ts
@@ -1,2 +1,5 @@
-export * from "./PostTable";
-export * from "./PostTableRow";
+export { default as PostTable } from "./PostTable";
+export { default as PostTableRow } from "./PostTableRow";
+export { default as PostAddDialog } from "./PostAddDialog";
+export { default as PostEditDialog } from "./PostEditDialog";
+export { default as PostDetailDialog } from "./PostDetailDialog";
